Move code-mode files
