+++ codex-rs/core/src/tools/code_mode/bridge.js
