+++ codex-rs/core/src/tools/handlers/mod.rs
@@ -1,7 +1,6 @@
 pub(crate) mod agent_jobs;
 pub mod apply_patch;
 mod artifacts;
-mod code_mode;
 mod dynamic;
 mod grep_files;
 mod js_repl;
@@ -32,10 +31,10 @@ use crate::function_tool::FunctionCallError;
 use crate::sandboxing::SandboxPermissions;
 use crate::sandboxing::merge_permission_profiles;
 use crate::sandboxing::normalize_additional_permissions;
+pub(crate) use crate::tools::code_mode::CodeModeHandler;
+pub(crate) use crate::tools::code_mode::CodeModeWaitHandler;
 pub use apply_patch::ApplyPatchHandler;
 pub use artifacts::ArtifactsHandler;
-pub use code_mode::CodeModeHandler;
-pub use code_mode::CodeModeWaitHandler;
 use codex_protocol::models::PermissionProfile;
 use codex_protocol::protocol::AskForApproval;
 pub use dynamic::DynamicToolHandler;
